requirements.txt
Django==5.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
cryptography==44.0.2
cffi==1.17.1
pycparser==2.22
python-dotenv==1.0.1
dj-database-url==2.2.0
whitenoise==6.7.0
psycopg[binary]==3.2.3